Rearranges interface
I rearranged some of the user interface because I didn't like the way it was set up. I simplified it and removed extraneous junk. I made the decision that users won't be able to define their own languages anymore. It means more work for me, but screw it, this is a side project, eh?
I will go in and fix the formatting in the things at a later time. Also to do at a later time, I must make block comment detection a real thing, and add support for more languages.